third_party/WebKit/LayoutTests/external/wpt/webvtt/api/VTTCue/lineAlign-expected.txt

This is a testharness.js-based test.
FAIL VTTCue.lineAlign, script-created cue assert_equals: expected (string) "start" but got (undefined) undefined
Harness: the test ran to completion.
